+++ tools/mtx_reader.c
@@ -5,7 +5,7 @@
 /* Reads a matrix in MartixMarket format, prints it in SMS format */
 
 int main() {
-	int prime = 42013;
+	i64 prime = 42013;
 	spasm_triplet *T;
 	
 	T = spasm_load_mm(stdin, prime);
